Added helpers_2

+++ printf.h
@@ -53,7 +53,7 @@ char			*set_case(char *str, t_flags flags);
 int				set_string(char **dst, char *src, int i);
 t_flags			init_flags(void);
 size_t			get_index(const char *str, char c);
-int				sep_calculator(int num);
+int				sep_calculator(long num);
 char			*set_separators(long num, t_length len, char *str, int start);
 
 #endif
